refactor: extract `kms` module out of `encryption` module
This should allow using KMS functionalities without the `basilisk` feature flag being enabled
